bot_rpg_quest/requirements.txt
pyTelegramBotAPI==4.14.0
python-dotenv